Rough draft of regression

+++ MLSandbox/src/NeuralNetwork.h
@@ -20,12 +20,12 @@ public:
 	inline Eigen::MatrixXf GetOutput() const { return m_CurrentOutput; }
 
 	void ForwardProp(Eigen::MatrixXf* input);
-	void BackwardProp(Eigen::VectorXi yTrue);
-	void Fit(Eigen::MatrixXf input, Eigen::VectorXi y, const Optimizer_SGD& optimizer);
+	void BackwardProp(Eigen::VectorXf yTrue);
+	void Fit(Eigen::MatrixXf input, Eigen::VectorXf y, const Optimizer_SGD& optimizer);
 
 	void Optimize(const Optimizer_SGD& optimizer);
-	float CalculateLoss(Eigen::VectorXi yTrue);
-	float CalculateAccuracy(Eigen::VectorXi yTrue);
+	float CalculateLoss(Eigen::VectorXf yTrue);
+	float CalculateAccuracy(Eigen::VectorXf yTrue);
 
 	Eigen::VectorXf GetQs(const Eigen::VectorXf& input);
 	int Predict(const Eigen::VectorXf& input);
